Check that input is a list in proper-list-length
proper-list-length and related functions claim to signal an error if the
input is not a proper list, and are careful to safely check that the cdr
is a cons (with safe-endp). However, they call cdr on the input without
checking that it is a list in the first place, resulting in memory-error
when compiled with safety 0 instead of type-error as the tests expect.